+++ src/writeCompressedPly.ts
@@ -3,6 +3,7 @@ import { FileHandle } from 'node:fs/promises';
 import { Splat } from './splat';
 import { Quat } from 'playcanvas';
 import { shNames } from './ply';
+import { generateOrdering } from './ordering';
 
 const generatedByString = `Generated by splat-transform ${version}`;
 const shBandCoeffs = [0, 3, 8, 15];
@@ -191,101 +192,6 @@ class Chunk {
     }
 }
 
-// sort the compressed indices into morton order
-const sortSplats = (splat: Splat, indices: Uint32Array) => {
-    // https://fgiesen.wordpress.com/2009/12/13/decoding-morton-codes/
-    const encodeMorton3 = (x: number, y: number, z: number) : number => {
-        const Part1By2 = (x: number) => {
-            x &= 0x000003ff;
-            x = (x ^ (x << 16)) & 0xff0000ff;
-            x = (x ^ (x <<  8)) & 0x0300f00f;
-            x = (x ^ (x <<  4)) & 0x030c30c3;
-            x = (x ^ (x <<  2)) & 0x09249249;
-            return x;
-        };
-
-        return (Part1By2(z) << 2) + (Part1By2(y) << 1) + Part1By2(x);
-    };
-
-    let minx: number;
-    let miny: number;
-    let minz: number;
-    let maxx: number;
-    let maxy: number;
-    let maxz: number;
-
-    const vertex = [0, 0, 0];
-    const it = splat.createIterator(['x', 'y', 'z'], vertex);
-
-    // calculate scene extents across all splats (using sort centers, because they're in world space)
-    for (let i = 0; i < indices.length; ++i) {
-        it(indices[i]);
-
-        const x = vertex[0];
-        const y = vertex[1];
-        const z = vertex[2];
-
-        if (minx === undefined) {
-            minx = maxx = x;
-            miny = maxy = y;
-            minz = maxz = z;
-        } else {
-            if (x < minx) minx = x; else if (x > maxx) maxx = x;
-            if (y < miny) miny = y; else if (y > maxy) maxy = y;
-            if (z < minz) minz = z; else if (z > maxz) maxz = z;
-        }
-    }
-
-    const xlen = maxx - minx;
-    const ylen = maxy - miny;
-    const zlen = maxz - minz;
-
-    if (xlen === 0 && ylen === 0 && zlen === 0) {
-        return;
-    }
-
-    const morton = new Uint32Array(indices.length);
-    let idx = 0;
-    for (let i = 0; i < indices.length; ++i) {
-        it(indices[i]);
-
-        const x = vertex[0];
-        const y = vertex[1];
-        const z = vertex[2];
-
-        // const ix = Math.min(1023, Math.floor(1024 * (x - minx) / xlen));
-        // const iy = Math.min(1023, Math.floor(1024 * (y - miny) / ylen));
-        // const iz = Math.min(1023, Math.floor(1024 * (z - minz) / zlen));
-
-        const ix = Math.floor(1024 * (x - minx) / xlen);
-        const iy = Math.floor(1024 * (y - miny) / ylen);
-        const iz = Math.floor(1024 * (z - minz) / zlen);
-
-        morton[idx++] = encodeMorton3(ix, iy, iz);
-    }
-
-    // sort indices by morton code
-    const mapping: Record<number, number> = {};
-    indices.forEach((v, i) => mapping[v] = i);
-    indices.sort((a, b) => morton[mapping[a]] - morton[mapping[b]]);
-
-    // sort the largest buckets recursively
-    // let start = 0;
-    // let end = 0;
-    // while (start < indices.length) {
-    //     while (end < indices.length && morton[sortIndices[start]] === morton[sortIndices[end]]) {
-    //         ++end;
-    //     }
-
-    //     if (end - start > 1) {
-    //         console.log('sorting', end - start);
-    //         sortSplats(splat, indices.subarray(start, end));
-    //     }
-
-    //     start = end;
-    // }
-};
-
 const chunkProps = [
     'min_x', 'min_y', 'min_z',
     'max_x', 'max_y', 'max_z',
@@ -342,8 +248,7 @@ const writeCompressedPly = async (fileHandle: FileHandle, splat: Splat) => {
 
     // sort splats into some kind of order (morton order rn)
     const sortIndices = new Uint32Array(numSplats).fill(0).map((_, i) => i);
-    sortSplats(splat, sortIndices);
-    indices.sort((a, b) => sortIndices[a.i] - sortIndices[b.i]);
+    generateOrdering(splat, sortIndices);
 
     const singleSplat = Chunk.members.map(_ => 0);
     const it = splat.createIterator(Chunk.members, singleSplat);
@@ -357,7 +262,7 @@ const writeCompressedPly = async (fileHandle: FileHandle, splat: Splat) => {
     for (let i = 0; i < numChunks; ++i) {
         const num = Math.min(numSplats, (i + 1) * 256) - i * 256;
         for (let j = 0; j < num; ++j) {
-            const index = indices[i * 256 + j];
+            const index = indices[sortIndices[i * 256 + j]];
 
             // read splat data
             it(index.i);
@@ -376,9 +281,9 @@ const writeCompressedPly = async (fileHandle: FileHandle, splat: Splat) => {
         }
 
         // repeat the last gaussian to fill the rest of the final chunk
-        // for (let j = num; j < 256; ++j) {
-        //     chunk.set(j, singleSplat);
-        // }
+        for (let j = num; j < 256; ++j) {
+            chunk.set(j, singleSplat);
+        }
 
         // pack the chunk
         chunk.pack();
